+++ src/components/Image/Image.tsx
@@ -12,12 +12,12 @@ export type ImageProps = {
   gradeBand?: GradeBand;
 };
 
-// const defaultImageSrc = '../../images/zelda.jpg';
+// const defaultImageSrc =
+//   import.meta.env.PROD === true && import.meta.env.VITE_ENV !== 'chromatic'
+//     ? '/node_modules/@connect/connect-design-system/dist/images/zelda.jpg'
+//     : '/images/zelda.jpg';
 
-const defaultImageSrc =
-  import.meta.env.PROD === true && import.meta.env.VITE_ENV !== 'chromatic'
-    ? '/node_modules/@connect/connect-design-system/dist/images/zelda.jpg'
-    : '/images/zelda.jpg';
+const defaultImageSrc = '/images/default.png';
 
 export const Image: React.FC<ImageProps> = ({
   imageSrc,