+++ src/index.ts
@@ -1,3 +1,4 @@
 import './assets/scss/custom.scss';
-
 export * from './components/';
+export * from './hooks';
+export * from './constants';