+++ src/types/opaque.ts
@@ -1,0 +1,1 @@
+export type Opaque<O extends string | symbol | number> = { __opaque__: O };